redis 구현에 따른 api 추가

+++ src/main/java/com/hgyr/multi/mainServer/data/Entity/User.java
@@ -55,7 +55,7 @@ public class User extends BaseEntity{
     @Column(nullable = false)
     @PositiveOrZero
     @Min(value = 0)
-    private Integer point;
+    private Double point;
 
     public User fromDto(UserDto userDto){
         this.uid = userDto.getUid();
